Admin: move DescribeLogDirs -> _cluster / cluster cli subgroup
