+++ kafka/cli/admin/cluster/__init__.py
@@ -1,6 +1,7 @@
 import sys
 
 from .describe import DescribeCluster
+from .log_dirs import DescribeLogDirs
 
 
 class ClusterSubCommand:
@@ -9,6 +10,6 @@ class ClusterSubCommand:
     def add_subparser(cls, subparsers):
         parser = subparsers.add_parser('cluster', help='Manage Kafka Cluster')
         commands = parser.add_subparsers()
-        for cmd in [DescribeCluster]:
+        for cmd in [DescribeCluster, DescribeLogDirs]:
             cmd.add_subparser(commands)
         parser.set_defaults(command=lambda *_args: parser.print_help() or sys.exit(2))
